+++ src/Field/Data/Story/StoryReferenceList.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace Torr\Storyblok\Field\Data\Story;
+
+use Torr\Storyblok\Field\Data\StoryReferenceData;
+
+final class StoryReferenceList
+{
+	/**
+	 * @param array<StoryReferenceData> $references
+	 */
+	public function __construct (
+		public readonly array $references,
+	) {}
+}